DoorLock .cpp, .h 파일 있음 삭제해주세요.
