+++ Embedded/Nucleo_mbed_template.uvprojx
@@ -134,11 +134,11 @@
             <RunIndependent>0</RunIndependent>
             <UpdateFlashBeforeDebugging>1</UpdateFlashBeforeDebugging>
             <Capability>1</Capability>
-            <DriverSelection>-1</DriverSelection>
+            <DriverSelection>4096</DriverSelection>
           </Flash1>
           <bUseTDR>1</bUseTDR>
           <Flash2>BIN\UL2CM3.DLL</Flash2>
-          <Flash3></Flash3>
+          <Flash3>"" ()</Flash3>
           <Flash4></Flash4>
           <pFcarmOut></pFcarmOut>
           <pFcarmGrp></pFcarmGrp>
@@ -2348,14 +2348,44 @@
               <FilePath>.\motordriver.h</FilePath>
             </File>
             <File>
-              <FileName>Doorlock.cpp</FileName>
+              <FileName>DHT22.cpp</FileName>
               <FileType>8</FileType>
-              <FilePath>.\Doorlock.cpp</FilePath>
+              <FilePath>.\DHT22.cpp</FilePath>
             </File>
             <File>
-              <FileName>Doorlock.h</FileName>
+              <FileName>DHT22.h</FileName>
               <FileType>5</FileType>
-              <FilePath>.\Doorlock.h</FilePath>
+              <FilePath>.\DHT22.h</FilePath>
+            </File>
+            <File>
+              <FileName>Adafruit_GFX.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>.\Adafruit_GFX\Adafruit_GFX.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>Adafruit_GFX.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\Adafruit_GFX\Adafruit_GFX.h</FilePath>
+            </File>
+            <File>
+              <FileName>Adafruit_GFX_Config.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\Adafruit_GFX\Adafruit_GFX_Config.h</FilePath>
+            </File>
+            <File>
+              <FileName>Adafruit_SSD1306.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>.\Adafruit_GFX\Adafruit_SSD1306.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>Adafruit_SSD1306.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\Adafruit_GFX\Adafruit_SSD1306.h</FilePath>
+            </File>
+            <File>
+              <FileName>glcdfont.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\Adafruit_GFX\glcdfont.h</FilePath>
             </File>
           </Files>
         </Group>
